moved submit tests
