+++ src/applications/health-care-questionnaire/shared/api/tests/submit-questionnaire-response/create.an.answer.array.unit.spec.js
@@ -1,6 +1,6 @@
 import { expect } from 'chai';
 
-import { createAnswerArray } from '../../../../shared/api/submit-questionnaire';
+import { createAnswerArray } from '../../submit-questionnaire';
 
 describe('health care questionnaire -- utils -- create an answer array', () => {
   it('empty array is created with a falsey answer', () => {
